requirements.txt
fastapi
uvicorn
websockets
python-dotenv
retell-sdk